todo

+++ src/Utils/ActionsMove.cpp
@@ -51,7 +51,7 @@ bool Move_CastSkillABC::UpdateMoveState(const DataPlayer &player_data, const Age
     static auto timer_wait = clock();
     static auto timer_casting = clock();
 
-    const auto skill_data = GW::SkillbarMgr::GetSkillConstantData((uint32_t)skill_cb->id);
+    const auto skill_data = GW::SkillbarMgr::GetSkillConstantData(skill_cb->id);
     if (!skill_data)
         return true;
     const auto cast_time_s = (skill_data->activation + skill_data->aftercast) * 1000.0F;
